% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_calculation.R
\name{calculate_complex_payments}
\alias{calculate_complex_payments}
\title{FUNCTION_TITLE}
\usage{
calculate_complex_payments(
  mortgage_length = 25,
  loan_principle = 2e+05,
  deposit = 10000,
  interest_vec = c(4.5, 5, 4.5),
  regular_payments = tibble::tibble(test = 100, test2 = 200),
  monthly_capacity = 2000,
  overpayment_chosen = 0,
  rental_income = 0,
  initial_sum_remaining = 0,
  initial_emergency_fund = 0,
  emergency_target = 0,
  savings_target = 0,
  saving_OP_pct = 0,
  monthly_outgoings = 0
)
}
\arguments{
\item{mortgage_length}{PARAM_DESCRIPTION, Default: 25}

\item{loan_principle}{PARAM_DESCRIPTION, Default: 2e+05}

\item{deposit}{PARAM_DESCRIPTION, Default: 10000}

\item{interest_vec}{PARAM_DESCRIPTION, Default: c(4.5, 5, 4.5)}

\item{regular_payments}{PARAM_DESCRIPTION, Default: tibble::tibble(test = 100, test2 = 200)}

\item{monthly_capacity}{PARAM_DESCRIPTION, Default: 2000}

\item{overpayment_chosen}{PARAM_DESCRIPTION, Default: 0}

\item{rental_income}{PARAM_DESCRIPTION, Default: 0}

\item{initial_sum_remaining}{PARAM_DESCRIPTION, Default: 0}

\item{initial_emergency_fund}{PARAM_DESCRIPTION, Default: 0}

\item{emergency_target}{PARAM_DESCRIPTION, Default: 0}

\item{savings_target}{PARAM_DESCRIPTION, Default: 0}

\item{saving_OP_pct}{PARAM_DESCRIPTION, Default: 0}

\item{monthly_outgoings}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[glue]{glue}}
}
