% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input_Costs.R
\name{calculate_initial_costs}
\alias{calculate_initial_costs}
\title{FUNCTION_TITLE}
\usage{
calculate_initial_costs(
  initial_costs_sum = 0,
  deposit = 0,
  total_liquid = 0,
  emergency_fund_target = 0
)
}
\arguments{
\item{initial_costs_sum}{PARAM_DESCRIPTION, Default: 0}

\item{deposit}{PARAM_DESCRIPTION, Default: 0}

\item{total_liquid}{PARAM_DESCRIPTION, Default: 0}

\item{emergency_fund_target}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
