% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_calculation.R
\name{calculate_mortgage}
\alias{calculate_mortgage}
\title{FUNCTION_TITLE}
\usage{
calculate_mortgage(
  df,
  interest_vec,
  emergency_target,
  savings_target,
  saving_OP_proportion,
  regular_payments
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{interest_vec}{PARAM_DESCRIPTION}

\item{emergency_target}{PARAM_DESCRIPTION}

\item{savings_target}{PARAM_DESCRIPTION}

\item{saving_OP_proportion}{PARAM_DESCRIPTION}

\item{regular_payments}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
