% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{count_decimals}
\alias{count_decimals}
\title{Count Decimals}
\usage{
count_decimals(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Vector with number of decimal places of each value in x
}
\description{
Get the number of decimal places from a numeric input
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
