% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{every_nth}
\alias{every_nth}
\title{Every nth}
\usage{
every_nth(n)
}
\arguments{
\item{n}{Value to return every n of}
}
\value{
Function
}
\description{
Return a function that returns every nth value from a vector, starting with the first. Used in printing axes in ggplot2
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
