% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{hr2}
\alias{hr2}
\title{hr tag with black line}
\usage{
hr2(style = "border-top: 1px solid #606060;", ...)
}
\arguments{
\item{style}{Style changes to `<hr>` tag, adjust this to change the style. Default: 'border-top: 1px solid #000000;'}

\item{...}{Parameters to be passed to `hr()`}
}
\value{
`hr()` shiny tag with a black line
}
\description{
Produce a `<hr>` tag with a solid black line, rather than grey.
}
\details{
#' Include Content From a File
#'
#' Load rendered RMarkdown from a file and turn into HTML.
#'
#' @rdname includeRMarkdown
#' @export
#'
#' @importFrom rmarkdown render
#' @importFrom markdown markdownToHTML
#' @importFrom shiny HTML
includeRMarkdown <- function(path){

  md <- tempfile(fileext = '.md')

  on.exit(unlink(md),add = TRUE)

  rmarkdown::render(
    path,
    output_format = 'md_document',
    output_dir = tempdir(),
    output_file = md,quiet = TRUE
    )

  html <- markdown::markdownToHTML(md, fragment.only = TRUE)

  Encoding(html) <- "UTF-8"

  return(HTML(html))
}

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
