% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output_TotalAssets.R
\name{plot_assets}
\alias{plot_assets}
\title{FUNCTION_TITLE}
\usage{
plot_assets(df, deposit, max_length = 25, initial_term = 3)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{deposit}{PARAM_DESCRIPTION}

\item{max_length}{PARAM_DESCRIPTION, Default: 25}

\item{initial_term}{PARAM_DESCRIPTION, Default: 3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate-joins}}
 \code{\link[tidyr]{pivot_longer}}
 \code{\link[tibble]{tibble}}
 \code{\link[scales]{label_dollar}}
}
