% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output_MonthlyHeadroom.R
\name{plot_headroom}
\alias{plot_headroom}
\title{FUNCTION_TITLE}
\usage{
plot_headroom(df, max_capacity = 0, disposable_spend = 0)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{max_capacity}{PARAM_DESCRIPTION, Default: 0}

\item{disposable_spend}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
