% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output_PaymentSplit.R
\name{plot_payment_split_by_month}
\alias{plot_payment_split_by_month}
\title{FUNCTION_TITLE}
\usage{
plot_payment_split_by_month(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
